joan: set stock firmware's system SPL as vendor SPL
This device tree currently sets PRODUCT_FULL_TREBLE_OVERRIDE, because
the vendor is capable of doing just that. This causes vendor security
patch level in Settings to show as Unknown, as we never actually *had*
a security patch due to being non-Treble on stock.

So set *a* vendor SPL to make it happy.

Change-Id: I3c51aa44086420980b621688373e9f6135ddf7c7